修改elementid baseParent为int64_t

+++ interfaces/innerkits/wm/window.h
@@ -1466,7 +1466,7 @@ public:
      * @param func Function to notify transfer component data.
     */
     virtual WMError TransferAccessibilityEvent(const Accessibility::AccessibilityEventInfo& info,
-        int32_t uiExtensionIdLevel) { return WMError::WM_OK; };
+        int64_t uiExtensionIdLevel) { return WMError::WM_OK; };
 
     /**
      * @brief Notify prepare to close window
